backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Mark",email = "wixw1cky7@gmail.com"}
]
requires-python = ">=3.13,<4"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "fastapi-limiter (>=0.2.0,<0.3.0)",
    "pyrate-limiter (>=4.0.2,<5.0.0)",
    "redis (>=7.3.0,<8.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "pyjwt[crypto] (>=2.11.0,<3.0.0)",
    "passlib[argon2] (>=1.7.4,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "apscheduler (>=3.11.2,<4.0.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "boto3 (>=1.42.65,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "faker (>=37.11.0,<38.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "ruff (>=0.14.1,<0.15.0)"
]
